Bug#20738072 CRASH ON STARTUP WITH MAX-DIGEST-LENGTH OF >=420K
Check for multiplication overflow before allocating Performance Schema arrays during server initialization.

+++ sql/sql_digest.h
@@ -29,7 +29,7 @@ class String;
 struct sql_digest_storage
 {
   bool m_full;
-  uint m_byte_count;
+  size_t m_byte_count;
   unsigned char m_md5[MD5_HASH_SIZE];
   /** Character set number. */
   uint m_charset_number;
@@ -82,8 +82,8 @@ struct sql_digest_storage
       as the thread producing the digest is executing concurrently,
       without any lock enforced.
     */
-    uint byte_count_copy= m_token_array_length < from->m_byte_count ?
-                          m_token_array_length : from->m_byte_count;
+    size_t byte_count_copy= m_token_array_length < from->m_byte_count ?
+                            m_token_array_length : from->m_byte_count;
 
     if (byte_count_copy > 0)
     {
